refactor: move to net.laprun.sustainability package / group id
