--- deployment/src/test/java/io/quarkiverse/power/deployment/TestPowerMeasurer.java
@@ -1,8 +1,0 @@
-package io.quarkiverse.power.deployment;
-
-import io.quarkus.test.Mock;
-import net.laprun.sustainability.power.PowerMeasurer;
-
-@Mock
-public class TestPowerMeasurer extends PowerMeasurer {
-}
