--- deployment/src/test/java/io/quarkiverse/power/deployment/TestPowerSensor.java
@@ -1,7 +1,0 @@
-package io.quarkiverse.power.deployment;
-
-import io.quarkus.test.Mock;
-
-@Mock
-public class TestPowerSensor extends net.laprun.sustainability.power.sensors.test.TestPowerSensor {
-}
