+++ deployment/src/test/java/net/laprun/sustainability/power/quarkus/deployment/PowerMeasurerTest.java
@@ -1,0 +1,30 @@
+package net.laprun.sustainability.power.quarkus.deployment;
+
+import io.quarkus.test.QuarkusUnitTest;
+import io.quarkus.test.common.http.TestHTTPResource;
+import net.laprun.sustainability.power.PowerResource;
+import net.laprun.sustainability.power.quarkus.runtime.PowerMeasurer;
+import net.laprun.sustainability.power.quarkus.runtime.ServerSampler;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.RegisterExtension;
+
+import java.net.URI;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class PowerMeasurerTest {
+    @TestHTTPResource
+    URI uri;
+
+    @RegisterExtension
+    static final QuarkusUnitTest config = new QuarkusUnitTest()
+            .withApplicationRoot((jar) -> jar.addClasses(PowerResource.class, TestPowerMeasurer.class, TestPowerSensor.class));
+
+    @Test
+    void startShouldAccumulateOverSpecifiedDurationAndStop() {
+        final var measurer = new PowerMeasurer(new ServerSampler(uri));
+
+        measurer.start(1, 100);
+        measurer.onCompleted(measure -> assertEquals(10, measure.numberOfSamples()));
+    }
+}
