+++ deployment/src/test/java/net/laprun/sustainability/power/quarkus/deployment/TestPowerMeasurer.java
@@ -1,0 +1,8 @@
+package net.laprun.sustainability.power.quarkus.deployment;
+
+import io.quarkus.test.Mock;
+import net.laprun.sustainability.power.PowerMeasurer;
+
+@Mock
+public class TestPowerMeasurer extends PowerMeasurer {
+}
