+++ deployment/src/test/java/net/laprun/sustainability/power/quarkus/deployment/TestPowerSensor.java
@@ -1,0 +1,7 @@
+package net.laprun.sustainability.power.quarkus.deployment;
+
+import io.quarkus.test.Mock;
+
+@Mock
+public class TestPowerSensor extends net.laprun.sustainability.power.sensors.test.TestPowerSensor {
+}
