--- integration-tests/src/test/java/io/quarkiverse/power/it/PowerResourceIT.java
@@ -1,7 +1,0 @@
-package io.quarkiverse.power.it;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class PowerResourceIT extends PowerResourceTest {
-}
