--- integration-tests/src/test/java/io/quarkiverse/power/it/PowerResourceTest.java
@@ -1,21 +1,0 @@
-package io.quarkiverse.power.it;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.Matchers.is;
-
-import org.junit.jupiter.api.Test;
-
-import io.quarkus.test.junit.QuarkusTest;
-
-@QuarkusTest
-public class PowerResourceTest {
-
-    @Test
-    public void testHelloEndpoint() {
-        given()
-                .when().get("/power")
-                .then()
-                .statusCode(200)
-                .body(is("Hello power"));
-    }
-}
