+++ integration-tests/src/test/java/net/laprun/sustainability/power/quarkus/it/PowerResourceIT.java
@@ -1,0 +1,7 @@
+package net.laprun.sustainability.power.quarkus.it;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class PowerResourceIT extends PowerResourceTest {
+}
