+++ pom.xml
@@ -7,7 +7,7 @@
     <artifactId>quarkiverse-parent</artifactId>
     <version>16</version>
   </parent>
-  <groupId>io.quarkiverse.power</groupId>
+  <groupId>net.laprun.sustainability</groupId>
   <artifactId>quarkus-power-parent</artifactId>
   <version>0.0.1-SNAPSHOT</version>
   <packaging>pom</packaging>
