+++ runtime/src/main/java/net/laprun/sustainability/power/quarkus/runtime/Sampler.java
@@ -1,0 +1,12 @@
+package net.laprun.sustainability.power.quarkus.runtime;
+
+import java.util.function.Consumer;
+
+public interface Sampler {
+
+    boolean isRunning();
+
+    void start(long durationInSeconds, long frequencyInMilliseconds) throws Exception;
+
+    void stop(Consumer<PowerMeasure> completed);
+}
